Integration Testing!
```
$ cargo test
```

now runs an integration test that uses the `gopro-chaptered-video-assembler`
binary to create real video output. A merkle tree is then constructed,
and the root node hash is compared with the expected output. I've
verified that the test fails when unexpected files are in either of the
directories, or the files don't match.

The reliance on extensive manual testing was making it harder to want to
work on this.

Fix #12